+++ src/main/java/nflx/rozhnov/transactionservice/model/Transaction.java
@@ -7,6 +7,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.UUID;
 
@@ -19,7 +20,7 @@ public class Transaction {
     @GeneratedValue
     private UUID id;
     private Date timestamp;
-    private long fromAccount;
-    private long toAccount;
-    private double amount;
+    private Long fromAccount;
+    private Long toAccount;
+    private BigDecimal amount;
 }
